# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Chat App API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ChatMessage {
  id: string;
  username: string;
  text: string;
  timestamp: string;
  imageUrl?: string;
}

export interface OnlineUser {
  username: string;
  socketId: string;
  joinedAt: string;
}

export interface DmMessage {
  id: string;
  from: string;
  to: string;
  text: string;
  timestamp: string;
  imageUrl?: string;
}

export type ListMessagesParams = {
  limit?: number;
};

export type GetDmHistoryParams = {
  /**
   * The requesting user's username
   */
  me: string;
  limit?: number;
};