# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Chat App API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /dm/{otherUsername}:
    get:
      operationId: getDmHistory
      tags: [dm]
      summary: Get DM history
      description: Returns direct message history between me and another user
      parameters:
        - name: otherUsername
          in: path
          required: true
          schema:
            type: string
        - name: me
          in: query
          required: true
          schema:
            type: string
          description: The requesting user's username
        - name: limit
          in: query
          schema:
            type: integer
            default: 100
      responses:
        "200":
          description: List of DMs
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/DmMessage"
components:
  schemas:
...
    DmMessage:
      type: object
      required: [id, from, to, text, timestamp]
      properties:
        id:
          type: string
        from:
          type: string
        to:
          type: string
        text:
          type: string
        timestamp:
          type: string
        imageUrl:
          type: string