# lib/api-zod/src/generated/types/dmMessage.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Chat App API specification
 * OpenAPI spec version: 0.1.0
 */

export interface DmMessage {
  id: string;
  from: string;
  to: string;
  text: string;
  timestamp: string;
  imageUrl?: string;
}